refactor(api): move method from repositories
we also copied findByUserId that is needed in the usecase so we can
avoid inject two repositories that shared the same name.
this method will probably be move to a shared context (or used behind an
api)
